Tentando resolver os endpoints
